refactor test file locations
