+++ test/service_manager_test.dart
@@ -9,7 +9,7 @@ import 'package:devtools/vm_service_wrapper.dart';
 import 'package:test/test.dart';
 import 'package:vm_service_lib/vm_service_lib.dart';
 
-import 'flutter_test_driver.dart';
+import 'support/flutter_test_driver.dart';
 
 void main() {
   group('serviceManagerTests', () {
